+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## CHANGELOG
 
+### UNRELEASED
+
+ - Fix additional properties for jsonschema conversion
+
 ### 0.14.0 - 2020-11-23
 
  - Improve to support schema with dynamic properties